Launcher3: Rewrite restart action for toggling settings
This implementation is much saner, doesn't abuse singletons and won't break in every literal merge
